feat: 新增盐和鱼两种资源，支持浮点资源数值
1. 新增盐、鱼两种城市资源配置与对应的建筑、图标资源
2. 将大部分资源数值类型从int改为float，优化数值精度展示
3. 调整结果面板、资源项的显示格式，支持浮点数值格式化展示
4. 更新AI相关逻辑适配浮点资源数值
5. 为部分城市添加盐、鱼资源的产出建筑与加成
6. 优化资源加成计算逻辑，支持城市产出加成

+++ Assets/Resources/Scripts/Controls/AI/AIToolHeroDev.cs
@@ -122,9 +122,9 @@ public static class AIToolHeroDev
             var attrConfig = CityAttrConfig.GetConfigByname(devCfg.DevAttr1.ToLower());
             if (!attrConfig.IsForceAttr)
             {
-                int currentVal = city.GetAttr(devCfg.DevAttr1);
+                float currentVal = city.GetAttr(devCfg.DevAttr1);
                 int maxVal = attrConfig.ValMaxCity;
-                int deficit = maxVal - currentVal;
+                float deficit = maxVal - currentVal;
                 if (deficit > 0)
                 {
                     score += deficit / 10f;
@@ -132,9 +132,9 @@ public static class AIToolHeroDev
             }
             else if (!attrConfig.IsPosRes)
             {
-                int currentVal = force.GetAttr(devCfg.DevAttr1);
+                float currentVal = force.GetAttr(devCfg.DevAttr1);
                 int maxVal = attrConfig.ValMaxForce;
-                int deficit = maxVal - currentVal;
+                float deficit = maxVal - currentVal;
                 if (deficit > 0)
                 {
                     score += deficit / 10f;
